+++ src/js/ui/tab-characters.js
@@ -10,6 +10,7 @@ const generics = require('../generics');
 const CharMaterialRenderer = require('../3D/renderers/CharMaterialRenderer');
 const M2Renderer = require('../3D/renderers/M2Renderer');
 const M2Exporter = require('../3D/exporters/M2Exporter');
+const CameraBounding = require('../3D/camera/CameraBounding');
 const WDCReader = require('../db/WDCReader');
 const ExportHelper = require('../casc/export-helper');
 const listfile = require('../casc/listfile');
@@ -383,7 +384,7 @@ async function previewModel(fileDataID) {
 
 		await activeRenderer.load();
 		//textureShaderMap = activeRenderer.shaderMap;
-		updateCameraBounding();
+		CameraBounding.fitObjectInView(renderGroup, camera, core.view.modelViewerContext.controls);
 
 		activeModel = fileDataID;
 
@@ -403,34 +404,6 @@ async function previewModel(fileDataID) {
 	core.view.isBusy--;
 }
 
-/** Update the camera to match render group bounding. */
-function updateCameraBounding() {
-	// Get the bounding box for the model.
-	const boundingBox = new THREE.Box3();
-	boundingBox.setFromObject(renderGroup);
-
-	// Calculate center point and size from bounding box.
-	const center = boundingBox.getCenter(new THREE.Vector3());
-	const size = boundingBox.getSize(new THREE.Vector3());
-
-	const maxDim = Math.max(size.x, size.y, size.z);
-	const fov = camera.fov * (Math.PI / 180);
-	const cameraZ = (Math.abs(maxDim / 4 * Math.tan(fov * 2))) * 6;
-
-	const heightOffset = maxDim * 0.7;
-	camera.position.set(center.x, heightOffset, cameraZ);
-
-	const minZ = boundingBox.min.z;
-	const cameraToFarEdge = (minZ < 0) ? -minZ + cameraZ : cameraZ - minZ;
-
-	camera.updateProjectionMatrix();
-
-	const controls = core.view.modelViewerContext.controls;
-	if (controls) {
-		controls.target = center;
-		controls.maxDistance = cameraToFarEdge * 2;
-	}
-}
 
 async function importCharacter() {
 	core.view.isBusy++;