+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Fixed issue that caused MH2O bitmaps in advanced data to be incorrect.
 - Fixed issue that preventing texture previews on the model viewer from working.
 - Fixed issue causing terrain UV maps to be non-normalized causing visible seams.
+- Fixed issue that caused model viewer camera to not properly show the model in frame.
 
 ### Internal
 - Updated three.js from r111 to r179 (some color changes might be observed in model previews).